requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
mlflow
joblib
dvc
